Check ActiveRecord::VERSION
