+++ lib/patches/active_record/relation.rb
@@ -1,10 +1,12 @@
-module Globalize
-  module Relation
-    def where_values_hash(relation_table_name = table_name)
-      return super unless respond_to?(:translations_table_name)
-      super.merge(super(translations_table_name))
+unless ::ActiveRecord::VERSION::STRING < "5.0.0"
+  module Globalize
+    module Relation
+      def where_values_hash(relation_table_name = table_name)
+        return super unless respond_to?(:translations_table_name)
+        super.merge(super(translations_table_name))
+      end
     end
   end
-end
 
-ActiveRecord::Relation.prepend Globalize::Relation
+  ActiveRecord::Relation.prepend Globalize::Relation
+end
